requirements.txt
discord.py
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
pandas~=1.2.4
matplotlib~=3.4.2
pafy~=0.5.5
youtube_dl
youtube-search


requests~=2.25.1